Change order of $(LFLAGS) in building to the last before -o for texpackr_test
As tested with ordering of flag for gcc, the order is important whenever
we try to statically link libpng in the executable. -lpng should be
first before -lz, but for static linking, we manually specify the
location of libpng16.a ourselves thus that should be placed before -lz
which is a system library.