Add parsing and meta functionality in texpackr_meta. Update Makefile.
It's a topic to discuss or think about it again later whether typedef
against struct is worth it. As now types.h has to include "hashmap_c.h"
in order to refer to its type. This probably increase compile time up
slightly with expense of user if never call relevant function related to
meta i.e. texpackr_parse(), or texpackr_sheetmeta_free().

We might revisit this to refactor the whole thing without typedef, and
explicitly use "struct" for every occurrences of using it.